feat: Add getter methods for number of time indices and ROIs
The code changes include the addition of getter methods `GetNumberOfTimeIndices()` and `GetNumberOfROIs()` in the `CCM` class. These methods return the number of time indices and regions of interest (ROIs) respectively. This change aims to provide easy access to these values for external use.

Note: The commit message follows the established convention of using a prefix to indicate the type of change (`feat` for feature addition).